pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hauba"
version = "0.1.0"
description = "Your AI Engineering Team — In One Command"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Hauba AI" },
]
keywords = ["ai", "agent", "llm", "automation", "multi-agent", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
]

dependencies = [
    "typer>=0.12,<1.0",
    "rich>=13.0,<14.0",
    "litellm>=1.40,<2.0",
    "aiosqlite>=0.20,<1.0",
    "httpx>=0.27,<1.0",
    "pydantic>=2.5,<3.0",
    "structlog>=24.1,<25.0",
    "pyyaml>=6.0,<7.0",
    "anyio>=4.0,<5.0",
]

[project.optional-dependencies]
computer-use = [
    "pyautogui>=0.9,<1.0",
    "playwright>=1.40,<2.0",
    "pillow>=10.0,<11.0",
]
voice = [
    "openai-whisper>=20231117",
    "edge-tts>=6.1,<7.0",
    "sounddevice>=0.4,<1.0",
]
web = [
    "fastapi>=0.110,<1.0",
    "uvicorn>=0.27,<1.0",
    "websockets>=12.0,<14.0",
]
channels = [
    "python-telegram-bot>=21.0,<22.0",
    "discord.py>=2.3,<3.0",
]
all = [
    "hauba[computer-use]",
    "hauba[voice]",
    "hauba[web]",
    "hauba[channels]",
]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "pytest-cov>=5.0,<6.0",
    "ruff>=0.5,<1.0",
    "pyright>=1.1,<2.0",
    "pre-commit>=3.7,<4.0",
]

[project.scripts]
hauba = "hauba.cli:app"

[project.urls]
Homepage = "https://github.com/NikeGunn/haubaa"
Documentation = "https://github.com/NikeGunn/haubaa#readme"
Repository = "https://github.com/NikeGunn/haubaa"
Issues = "https://github.com/NikeGunn/haubaa/issues"
Changelog = "https://github.com/NikeGunn/haubaa/releases"

[tool.hatch.build.targets.wheel]
packages = ["src/hauba"]

[tool.hatch.build.targets.sdist]
include = [
    "src/hauba/",
    "tests/",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

[tool.hatch.build]
artifacts = [
    "src/hauba/bundled_skills/*.md",
    "src/hauba/bundled_strategies/*.yaml",
    "src/hauba/py.typed",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",    # Line too long (handled by formatter)
    "SIM105",  # contextlib.suppress style preference
    "SIM102",  # Nested if style preference
    "SIM115",  # Context manager for file open style
    "RUF006",  # asyncio.create_task reference storage
    "B904",    # raise from within except
    "UP042",   # StrEnum (Python 3.11+ only)
    "SIM117",  # Nested with statements style preference
    "RUF043",  # Regex metacharacters in pytest match
]

[tool.ruff.lint.isort]
known-first-party = ["hauba"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
include = ["src"]
exclude = ["tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: tests requiring file system and SQLite",
    "e2e: end-to-end CLI tests",
]
